Avoid dependency on time-sensitive cache in ZoneInfoDb from TimeZone/ZoneInfo
Instead, the cache in ZoneInfoDb becomes time-insensitive, and
the DST savings can be obtained from the new API
ZoneInfoData.getLatestDstSavings and specifying the time.

The expected behavior change is that the public API
TimeZone.useDaylightTime() and TimeZone.getDSTSavings() will use
the system clock when ZoneInfo instance is created, instead of the
arbitrary current clock time used in the cache.

This change helps fix an edge case of incorrect DST savings from
the public APIs when
- a time zone transits from using DST to using a new DST/not using it, and
- the device/app has not rebooted since the transition because
  Zygote/app process has the outdated cache.

Add testUseDaylightTime_consistency() to verify no behavior change
i.e. mUstDst is true only if mDstSavings != 0 before and after this change.

Bug: 172654324
Bug: 139480281
Test: atest CtsLibcoreTestCases
Change-Id: I794c8f50b5e377f304e04e691282abea0a29bf13