Deflake OsTest.test_socket_setSockoptTimeval_effective.
Be more tolerant of longer delays for when device is under load.

Additionally check that the process blocks for *at least* the
expected timeout.  This is a potential source of new flakes
but the socket is bound but not listening, so read should
block forever in the absence of a timeout.

Hans, you've tweaked similar timeouts in the past, do these
seem about right to you?

Bug: 175252827
Test: atest OsTest.test_socket_setSockoptTimeval_effective.
Change-Id: I615f14b6d6b7e3e3e6d7a5a23f579d3681fad581