Add script to help with merging from upstream
Added a script which makes use of Git's 3-way merge conflict
resolution to update ojluni files to a newer version from the
upstream.

The tool is primarily useful to get the chance to experiment with
newer versions and extract information on the necessary work for
a successful migration. Manual inspection of all the updated
files may still be necessary, as the automatic conflict
resolution may have missed something.

Test: merge-from-upstream -t 11+28 java/util/concurrent
Bug: 188889082
Change-Id: I2d9aa066ba5a428a30bc2dd339afc251614494d9