Allow Connectivity usage of core-tests-support
Connectivity test suites are being reorganized to add coverage tests,
and core-tests-support needs to be visible from
Connectivity/tests/unit instead of the previous path.

Allow all Connectivity test subpackages to use core-tests-support,
so that the visibility does not need to be updated when test suites
are reorganized inside Connectivity.

Bug: 187935317
Test: m
Change-Id: I5fbc205c31c52ff08db834fdc8dda6dd7f06e31f