Suppress a few lint errors in libcore
Because r.android.com/1689750 is not in API 31, lint now detects
use of these newly-public fields are only available on 32+, and thinks
access to them requires API checks within libcore code. This is
incorrect, so suppress the errors.

If the CL references above is reverted, this CL should be revertable
as well.

Bug: 171506470
Bug: 183513856
Test: m out/soong/.intermediates/libcore/core-libart/android_common/lint/lint-report.txt
Change-Id: I2a114de952e6175a673508d6f63e8ac053039e18