Revert an unnecessary patch in java.text.CollationElementIterator
The patch is unnecessary because NULLORDER is constant and inlineable.
from the public Android SDK, as shown in api/current.txt,
we can't change the value anyway. So assigning the value from ICU4J
CollationElementIterator seems unnecessary.

The goal here is actually to remove the static initialization block
in legacy.art.module.platform.api.stubs, and thus the stub will be
identical to art.module.public.api.stubs.module_lib

Bug: 192446466
Test: Verified that the class file size are identical.
      diff <(unzip -l out/soong/.intermediates/libcore/mmodules/core_platform_api/legacy.art.module.platform.api.stubs/android_common/turbine-combined/legacy.art.module.platform.api.stubs.jar) <( unzip -l out/soong/.intermediates/libcore/art.module.public.api.stubs.module_lib/android_common/turbine-combined/art.module.public.api.stubs.module_lib.jar)
Change-Id: I2902963704ddcb6d8c1583dc5165f2e4da1a9e3c