MappedByteBuffer: use sizeof(*val) instead of sizeof(T)
The static analyzer has complaints for `(T *)malloc(sizeof(V))`, and
_technically_ `unsigned char` != `char`. Hence, this code was upsetting
the analyzer.

The simplest way to fix this is to use the
`T *x = (T *)malloc(N * sizeof(*x));` pattern instead of repeating `T`
in the `sizeof`. These are equivalent, but the latter is more resilient
to refactors.

Caught by the static analyzer:
> libcore/ojluni/src/main/native/MappedByteBuffer.c:53:43: warning:
Result of 'malloc' is converted to a pointer of type 'unsigned char',
which is incompatible with sizeof operand type 'char'
[clang-analyzer-unix.MallocSizeof]

Bug: None
Test: TreeHugger

Change-Id: Id424b7c349bed433e5984b3f73045f2cc6709dca