Remove @CorePlatformApi usage in libcore/
No behavior change is expected, because the kCorePlatformApi bit of
each method / field is still set in the core-libart and core-oj dex
files. During the build of ART module, the module-lib stub is passed as
a core_platform_api stub in the "art-bootclasspath-fragment", and thus set
the kCorePlatformApi into the generated dex files.

Bug: 192446466
Test: m droid
Change-Id: Ia9371b93184c678dc42f2bff582f381d1e5f5ac2