Switch hwasan from fork to vfork.
Remove the special case for "hwaddress" sanitizer (HWASan) that makes it
use fork instead of vfork (which is the default). The case was
introduced earlier because of the missing support for vfork in HWASan.

HWASan has been fixed more than 2 years ago in
https://reviews.llvm.org/D58533. Moreover, a bug in the fork support has
been discovered and fixed upstream in https://reviews.llvm.org/D108063,
but it is not in the current Android toolchain yet.

Switching back to vfork makes the code faster, avoids the hwasan bug,
and reduces the diff in behavior between hwasan and regular builds.

Bug: 196623668
Test: CtsLibcoreTestCases (no new failures)
Test: manual
Change-Id: I15d32829806f605d5599d776753bc32df0494661