Using git rerere in the merge from upstream tool
Making use of the git rerere tool to cache resolution information
when merging in changes from upstream.

The use is optional and switched off by default. It needs a cache
folder where all the resolutions are stored. The script will copy
the cache folder in the work git dir. This will help git merge to
apply same resolutions.

Test: merge-from-upstream -t 11+28 java/util/concurrent
Test: merge-from-upstream -t 11+28 -r /tmp/cache-dir java/util/concurrent
Bug: 188889082
Change-Id: Iafcf3b0cdd93719d8d2a76ba39630203fd936d63