[DO NOT MERGE] CTS 11: Skip XDH if present in KeyFactory2Test.
Part of a set of changes to Android 11 CTS to enable it to pass
with or without an Android 12-based Conscrypt module installed.

XDH KeyFactory may or may not be present depending on which
Conscrypt module is installed. Android 11 CTS should be agnostic
to its presence, so skip it during tests.

Bug: 195615724
Test: atest CtsLibcoreTestCases with and without a 12-based Conscrypt module
Change-Id: Ic2107e5641dfbaf90e5be512c03ab0ca2c4c985a