[DO NOT MERGE] CTS 10: Skip XDH if present in KeyFactory2Test.
Part of a set of changes to Android 10 CTS to enable it to pass
with or without an Android 12-based Conscrypt module installed.

This change is a backport of the equivalent CTS 11 change
in https://r.android.com/1832455

XDH KeyFactory may or may not be present depending on which
Conscrypt module is installed. Android 10 CTS should be agnostic
to its presence, so skip it during tests.

Bug: 197500794
Bug: 195615724
Test: atest CtsLibcoreTestCases with and without a 12-based Conscrypt module
Change-Id: Ic2107e5641dfbaf90e5be512c03ab0ca2c4c985a
(cherry picked from commit 87c546e45d48471b57da79767c73bd23b3b60bae)