Fixing j.u.c.ScheduledExecutorTest(s)
The ScheduledThreadPoolExecutor has been updated to handle
delayed tasks differently when the executor is cancelled. After
the update, the expired tasks do not get cancelled when the
executor onShutdown() is called.

The tests have been updated to make sure that delayed tasks are
not expired by the time cancel is called.

Test: atest jsr166.ScheduledExecutorSubclassTest#testShutdown_cancellation
Test: atest jsr166.ScheduledExecutorTest#testShutdown_cancellation
Test: atest CtsLibcoreOjTestCases:ScheduledExecutorTest#testShutdown_cancellation
Test: atest CtsLibcoreOjTestCases:ScheduledExecutorSubclassTest#testShutdown_cancellation
Bug: 188889082
Bug: 202927404

Change-Id: Ia03e262ef17ce6c7f3ef1b3226b9a0b9b2878bc7