Adding MethodHandles.privateLookupIn
This is needed by OpenJDK 11+28 version of
java.util.concurrent.atomic.Striped64 for accessing a member
variable in java.lang.Thread.

There is no module logic in Android, so this is being bypassed.
Also there is no security manager and the function cannot be
called in privileged mode. For this reason, java.lang.Thread is
added to the allowed lookup list.

Test: m
Test: atest CtsLibcoreTestCases:libcore.java.lang.invoke.MethodHandlesTest
Bug: 190468875
Change-Id: I550ec5cdd0d84a94ca98443416d031d8b7112ddd