Move ObjectStreamClassTest into dedicated Java 9 suite
With the move to Java 11, the ObjectStreamClassTest will stop
working due to the way Java 11 handles nested classes. The
private inner classes used in the test will generate a different
SUID from the pre-11 version. We cannot preserve this
compatibility when moving to Java 11, however the behaviour of
the core libraries must be preserved for any prior versions.

The test has been moved in order to force Java 1.9 compilation
for these classes.

Test: atest -m core-tests:libcore.java.io.ObjectStreamClassTest
Bug: 211750879
Change-Id: Ia2b3c135eb8dc3c98685c49c495dddd057a36cd7