Add ObjectStreamClassSuidTest
As ObjectStreamClassTest was moved in order to be compiled with
Java 1.9, a similar test has been added to check that computed
SUIDs are preserved by the core libraries. The only difference
from the original code is that the inner classes are no longer
private. This ensures that there is no need to generate the inner
token object to preserve the private access.

Test: atest -m core-tests:libcore.java.io.ObjectStreamClassSuidTest
Bug: 211750879
Change-Id: I1e9ed2fced7d8259043f8515c8c8c0d4f030a7a9