Add API tests for FileStore.
FileStore itself is an abstract class, with currently
no API test coverage.  To test it this CL pulls in the
OpenJDK11 implementation of UnixFileStore which contains the
new API we need to test.  Testing is then via the existing
LinuxFileSystemTest, and is fairly mimimal as the API surface
is small and simple.

Fixes: 203426354
Test: atest CtsLibcoreTestCases:libcore.java.nio.file.LinuxFileSystemTest
Change-Id: Ic5ff758c3775154185b904a1b96470a6a0fa613f