Clean stale Locale objects in Zygote before forking
In Zygote, as part of the initialization, there are many `BaseLocale`
and `Locale` objects being cached with SoftReference
(see LocaleObjectCache.java).  Many of these cached objects will be
garbage collected, and their SoftReference entries in the
LocaleObjectCache need to be cleaned.  Unfortunately, the cleaning
of the stale entries happens in every child process from Zygote
upon the first use of Locale, which is causing dirty pages (the
cleaning happens in each child process's separate address space),
as well as wasting cpu cycles doing the same work.

This CL fixes this issue by explicitly triggering the cleaning of
the stale entries cached after GC but before Zygote forks.

See below bug for more details.

Bug: b/204785359

Change-Id: Iacda3a6c106576fed50a1418982aa43ebec59ee7