Fix flake in ExplicitCastArgumentsTest#testMultipleArg
Update non-null unboxing test to use the type of the object instance
rather than the type from the MethodType. This allows boxed
primitives, for example Integer instances, to be unboxed when they
appear as Object.class in the MethodType.

This change also increases the default number of iterations of the
failing test since it uses a random number generator to create
inputs. Previously it was failing within 1-10 iterations.

Bug: 217356401
Test: art/test.py --host
Test: atest --rerun-until-failure 1000 \
  CtsLibcoreOjTestCases:test.java.lang.invoke.ExplicitCastArgumentsTest#testMultipleArgs
Test: atest CtsLibcoreOjTestCases:test.java.lang.invoke
Test: CtsLibcoreTestCases:libcore.java.lang.invoke
Change-Id: I74081d8b79ca4a6d14e17f302152345f899bade0