Revert^2 "More consistently use refersTo()"
This reverts commit e9765f29e97d55942863d135f62cbe68097be82a.

PS1 is identical to aosp/1964279.
PS2 rolls reverts one of the PS1 changes in ThreadLocal.

That change was not correct, since it introduced a possibility of
an unexpected null, under rare timing conditions. On closer look,
it also does not appear very useful, for the same reason that we
originally left WeakHashMap alone: The referent is either already
null, or we need the result of .get() anyway. None of the other
changes have this issue.

Test: Build and boot AOSP
Bug: 215259836
Bug: 217171086

Change-Id: I803988122cc16e7eaf9dcfa584b02489758c0929