Add API tests for FileStore.
This change re-lands https://r.android.com/1966880
with changes to make it more tolerant for host testing.

FileStore itself is an abstract class, with currently
no API test coverage.  To test it this CL pulls in the
OpenJDK11 implementation of UnixFileStore which contains the
new API we need to test.  Testing is then via the existing
LinuxFileSystemTest, and is fairly mimimal as the API surface
is small and simple.

Fixes: 203426354
Test: atest
CtsLibcoreTestCases: libcore.java.nio.file.LinuxFileSystemTest
Change-Id: Id7d1412620d1ac256660a113c82a5361d6db4fde