Enable TestZoneRules#negativeDST
The test failed, but has nothing to do with negative DST.

During 1969 - 1971 summers, Europe/Dublin is using GMT+1 as the standard
time, and using GMT+1 since 27-Oct-1968 until 31-Oct-1971.

Thus, this CL changes the test expectation for Europe/Dublin, and
enables the test.

Sources:
1. "europe" file in the IANA time zone data
=======================
 # Zone  NAME            STDOFF  RULES   FORMAT  [UNTIL]
 Zone    Europe/Dublin    0:00   GB-Eire GMT/IST 1968 Oct 27
 # Vanguard section, for zic and other parsers that support negative DST.
                         1:00   Eire    IST/GMT
 # Rearguard section, for parsers lacking negative DST; see ziguard.awk.
 #                        1:00   -       IST     1971 Oct 31  2:00u
========================
No matter it's using vanguard / rearguard data, the standart time should
be GMT+1. So the test should expect the standard time GMT+1 no matter
negative time is supported or not.

2. Wikipedia
https://en.wikipedia.org/wiki/Time_in_the_Republic_of_Ireland
In Ireland, the Standard Time Act 1968 legally established that the time for general purposes in the State (to be known as standard time) shall be one hour in advance of Greenwich mean time throughout the year.[2] This act was amended by the Standard Time (Amendment) Act 1971, which legally established Greenwich Mean Time as a winter time period.[1]

Bug: 180577079
Test: atest CtsLibcoreOjTestCases:test.java.time.zone.TestZoneRules
Change-Id: Ia321bceecde9de738ccfffdba90e76cdd400bac9