Fix DdmVmInternalTest
This test was added primarily for test coverage and it does not have a
lot of scence to assert two profiles' length for equality when
setRecentAllocationsTrackingEnabled() is set to false and true.

This CL removes this assertion and makes test always pass. Should still
add the necessary method coverage for DdmVmInternal class.

Bug: 227676094
Test: atest libcore.org.apache.harmony.dalvik.ddmc.DdmVmInternalTest
Change-Id: I4c55c38469e3f2d8f0ec8b31b960f7ca42beaf13