Soft removal of Runtime.getLocalizedInput/OutputStream API
The API has been a no-op. The API has been deprecated since API level 14,
and removed from OpenJDK 11. It seems to be the best candidate
for trying soft API removal process in libcore and ART mainline module.

Existing app should still be able to call the API, but the API will
be removed in the next SDK release.

Bug: 254453420
Test: m droid
Ignore-AOSP-First: Merge Conflict with internal master
Change-Id: I48dd3fdf050624d65ce682e8632441c0d4e7e8b6