Add a script to run checkstyle and use it in PREUPLOAD.cfg
This change adds a script `libcore-checkstyle.sh` that applies
checkstyle to different sub-projects within libcore. The script is
intended to run as a repohook, but can also be run standalone for
testing.

The repohook is updated to invoke `libcore-checkstyle.sh`.

This change also adds an AOSP copyright check. It also fixes a couple
of AOSP files that were missing a period in the header copyright.

Bug: 255273691
Test: tools/checkstyle/libcore-checkstyle.sh
Test: manually add a successor commit dropping copyright headers
Change-Id: If2347489067906764a9f14e11427953120d4eefb