Add j.a.p.Generated to ART's Public API surface
Certain code generation tools like Dagger emit this annotation. Add this
to Public API surface to enable apps to use these annotation processors.

Another option was to add this to the new "Toolchain" API surface
instead. e.g. LambdaMetaFactory (LMF) was added to Toolchain API surface
and not Public API in aosp/2369873. One difference is that unlike LMF,
j.a.p.Generated is part of JLS and is backed by an implementation.

Bug: 261244752
API-Coverage-Bug: 266480376
Test: TH

Change-Id: I29122adc5ee9d7cafd5403ac5421a7b0d9737dd1