json: Improve JSON deserialization performance
Since JDK 8 String#substring no longer contains a reference to the
character array of the originating String thus eliminating the
possibility of memory leaks. The call to new String immediately after
substring simply slows things down. The call to String#length is not
free since it involves invoking a function and performing a shift
operation so instead cache the value and reuse it in tight loops.

Test: atest CtsLibcoreTestCases

Change-Id: I9d3684b660e22abe287a21d8817d2f859a136399
Signed-off-by: Jacob Abrams <satur9nine@gmail.com>