Update CertificateTest.
Depending on the version of the Conscrypt module in use, MD5-signed
certificates may be rejected by the CertPath APIs.  MD5 tests will
be skipped if not supported, so this test should be agnostic about
Conscrypt version.

Also added tests for SHA1-signed certificates.

These certificate types are rejected by Android's TrustManager but
need to be supported by the public CertPath APIs, at least as
long as BoringSSL supports them.

Bug: 270301670
Test: atest CtsLibcoreTestCases:tests.targets.security.cert.CertificateTest
Change-Id: If2a4739cd7b89e65136e5e2b50d1af810930ee6f