Serialize and de-serialize record objects
It's not possible to upgrade to entire ObjectInputStream / OutputStream
to jdk17u because it involves new other features, e.g.
ObjectInputFilter.

This CL integrates the de- / serialization codes for record type
from jdk-17.0.5-ga. Upstream has a couple of commits for this, and thus,
it's not possible to do a cherry-pick. Several partial cherry-pick of
commits will be needed but it's more confusing than the current
approach.

I only added change marker Android-changed for the manual change I made.

Bug: 272698028
Test: atest CtsLibcoreTestCases:crossvmtest
Test: atest CtsLibcoreOjTestCases:test.java.io.Serializable.records
Change-Id: I1b6f6dac33996761541a5c09bf6a30f9ffcc23a2