Refine isZipPathValidatorEnabled for ZipFile
aosp/2578354 introduced isZipPathValidatorEnabled instance variable to
ZipFile which optimises loading APKs. In order to further optimise zip
path validation, this change adds an extra check around
ZipPathValidator's callback: if the callback is the default no-op
implementation, then zip path validation is not required. This removes
the need to do java calls to onZipEntryAccess() if the callback is a
no-op.

Bug: 272230610
Test: atest CtsLibcoreTestCases:BaseDexClassLoaderTest
Test: atest CtsLibcoreTestCases:libcore.libcore.io.ClassPathURLStreamHandlerTest
Test: atest JarFileTest
Test: atest CtsLibcoreTestCases:libcore.java.util.zip.ZipFileTest
Test: atest CtsSecurityHostTestCases:android.security.cts.host.ZipPathValidatorTest
Change-Id: I650728d0fdb7911e23596042420d016394d649ec