Update API due to changes in metalava
Metalava changes in this topic have added some previously missing
`toString()` methods to some final classes.

The `toString()` method was implemented by the `ChronoLocalDateImpl`
class which is not part of the API, and defined in the
`ChronoLocalDate` interface which is part of the API.

It is not clear exactly why this change happened but it is likely due
to changes in the version of intellij jars and its handling of implicit
modifiers.

Bug: 294300564
Bug: 281035030
Test: ./gradlew
      m update-api
(cherry picked from https://android-review.googlesource.com/q/commit:88ccf8244c1c889636bfdbb682805db446653877)
Merged-In: I06248d7ef77a46d084e78713efa02299c19236fb
Change-Id: I06248d7ef77a46d084e78713efa02299c19236fb