Fix a native call to ZipFile#onZipEntryAccess()
The 2nd parameter of ZipFile#onZipEntryAccess() is a flag but the native
code that calls that method passes the length of the entry string
instead which is wrong. This change fixes this issue by passing the
flag as the 2nd parameter.

Bug: 274918229
Test: atest CtsLibcoreTestCases:libcore.java.util.zip.ZipFileTest
Test: atest CtsSecurityHostTestCases:android.security.cts.host.ZipPathValidatorTest
Test: atest CtsLibcoreTestCases:libcore.dalvik.system.ZipPathValidatorTest
Merged-In: Ifc567c581d31fed7e579bb29c00fd05246805d28
Change-Id: I5e397a70560275657897430f99dbdad9e45e7b54