Fix a native call to ZipFile#onZipEntryAccess()
This change fixes an obvious issue so seems relatively safe. We want to
merge this first because we're past the ZBB-P1, but there's already an
AOSP cherry pick aosp/2607772.

The 2nd parameter of ZipFile#onZipEntryAccess() is a flag but the native
code that calls that method passes the length of the entry string
instead which is wrong. This change fixes this issue by passing the
flag as the 2nd parameter.

Bug: 274918229
Test: atest CtsLibcoreTestCases:libcore.java.util.zip.ZipFileTest
Test: atest CtsSecurityHostTestCases:android.security.cts.host.ZipPathValidatorTest
Test: atest CtsLibcoreTestCases:libcore.dalvik.system.ZipPathValidatorTest
Ignore-AOSP-First: Past ZBB-P1, will merge to AOSP later
Change-Id: Ifc567c581d31fed7e579bb29c00fd05246805d28