Better expose exceptions in cleaners/finalizers
Do not catch exceptions in SystemCleaner Runnables, thus allowing
them to crash the process. The RI spec promises that normal Cleaners
ignore exceptions, as finalizers did. That potentially hides
errors. We cannot change the behavior of normal Cleaners without
introducing compatibility issues. But since we control SystemCleaner,
have it exhibit what we believe is better behavior.

Delay timeouts while a finalizer is trying to log an uncaught
exception, so that we have a chance of identifying the cause.

Avoid an NPE if we time out while the daemon doesn't appear to be
working on anything. This can happen if the daemon crashes without
killing the entire process, which is easily possible in a test
environment.

Bug: 282377285
Test: TreeHugger
Change-Id: Ibbfb3ad8cd4ece1dc668f13510a950f3ef3cbcc7