Update notices-for-stubs.jar only if it is changed
notices-for-stubs.jar triggers most apk and jar files to be rebuilt,
which takes much of build time.
However, there are some cases that touch the notices-for-stubs.jar
without any changes of the notices-for-stubs.jar.
To avoid this waste of time, add `write_if_changed: true` option to
update the notices-for-stubs.jar file only if genrule changed the
file.

Bug: 290130959
Test: See if build time increase much by lunching and building a
different target

Change-Id: I3508afb86daab391f5a059eec8d04d80ac3a14f7