Add IPV6_PKTINFO constant in OsConstants
IPV6_PKTINFO socket option allows user space to specify the IPv6
source address and interface for an outgoing packet with ancillary
data on sendmsg(), or retrieve these information from a received
packet on recvmsg(). This socket option is useful in situations
where we have to specify the desired global IPv6 address when sending
the packets.

Bug: 293393743
Test: TH
Change-Id: I3a3029a286f345e1e4df4abf2cf72e4840a9c0cb