Remove unusable libcore.net.NetworkSecurityPolicy constructor
Enabling a metalava API check that was previously disabled for the intra
core API revealed a problem with the `NetworkSecurityPolicy`. In the
`@IntraCoreApi` surface it has a public constructor but the method
`isCleartextTrafficPermitted()` is abstract and hidden. Any class that
attempted to subclass `NetworkSecurityPolicy` through the `@IntraCoreApi`
surface would end up with a class that would not work at runtime because
it would be missing an implementation of `isCleartextTrafficPermitted()`.

That is not an issue when subclassing it via the `@SystemApi` surface as
`isCleartextTrafficPermitted()` is included in that API surface.

This change removes the public constructor from the `@IntraCoreApi`
surface. Although, that change is theoretically an incompatible change it
is safe because:
* There can be no working subclasses of `NetworkSecurityPolicy` created
  via the `@IntraCodeApi`.
* The intra core api surface of the `NetworkSecurityPolicy` class is
  only and has ever only be used by conscrypt and it has only ever used
  two methods, `getInstance()` and
  `isCertificateTransparencyVerificationRequired`.

Bug: 292484829
Test: ./gradlew
Change-Id: I5c43e691cb0442bb4ffd431242b467bb22cbd4f3