Enable more asserts in LegacyLocalesTest.
Initially I thought that ICU.getAvailableLocales() was initialized
in Zygote, but it is not - obsoleteLocales_withFlagEnabled was
failing due inconsistent state of USE_NEW_ISO_LOCALE_CODES at
caches initialization and test run times. Cleaning caches before
tests run to make sure that flag's state is the same both when
caches are populated and when test itself is executed.

It relies heavily on the ICU implementation details as there are
no APIs to reset caches.

Bug: 291868760
Test: atest CtsLibcoreTestCases
Change-Id: I2abc5d6a7a42471b46444fbf2f61f91ebe762ae2