Add some missing overrides in EnumMap and DelayQueue
Previously, the `EnumMap.put(K,V)` method was excluded from the API
signature file because Metalava thought it was the same as
`AbstractMap.put(K,V)`. While they certainly look the same the two
`K`s are not the same because they have different bounds:
* In `AbstractMap<K,V>` the `K` extends `Object` implicitly.
* In `EnumMap<K extends Enum<K>>` the `K` extends `Enum` explicitly.

So, when taking into account the type parameter bounds (which Metalava
did not do before) the two methods are not the same. That can easily be
seen when looking at the erased types, `EnumMap.put(Enum,Object)` and
`AbstractMap.put(Object,Object)`.

The Metalava issue has been fixed and it now takes type parameter
bounds into account which means that the `EnumMap.put(K,V)` is retained
in the signature file.

A similar argument applies to why `DelayQueue.add(E)` is now also kept.

Bug: 306245305
Test: m checkapi
Change-Id: I924741db5f617e01c180ef9330a31e614a926cd8